feat: optimize elo updates with upsert function
