+++ op_tcg/backend/models/leader.py
@@ -1,3 +1,4 @@
+from datetime import date
 from enum import StrEnum
 
 from pydantic import Field
@@ -69,3 +70,14 @@ class TournamentWinner(BQTableBaseModel):
     meta_format: MetaFormat = Field(description="Meta in which tournaments happened", primary_key=True)
     leader_id: str = Field(description="The op tcg leader id e.g. OP03-099", primary_key=True)
     count: int = Field(description="Number of tournament wins")
+
+
+class LeaderElo(BQTableBaseModel):
+    _dataset_id: str = BQDataset.LEADERS
+    meta_format: MetaFormat = Field(description="Meta until or in which the elo is calculated", primary_key=True)
+    leader_id: str = Field(description="The op tcg leader id e.g. OP03-099", primary_key=True)
+    only_official: bool = Field(default=False, description="Whether the matches are only originated from "
+                                                           "official tournaments", primary_key=True)
+    elo: int = Field(description="Elo rating of leader until a certain time/ meta format")
+    start_date: date = Field(description="Date in which the elo calculation started")
+    end_date: date = Field(description="Date in which the elo calculation ended")
