+++ op_tcg/backend/models/matches.py
@@ -1,13 +1,13 @@
 import pandas as pd
-import pandera as pa
 from pydantic import BaseModel, Field
-from datetime import datetime, date
+from datetime import datetime
 from enum import IntEnum
 
 from op_tcg.backend.models.bq_enums import BQDataset
 from op_tcg.backend.models.bq_classes import BQTableBaseModel
 from op_tcg.backend.models.common import DataSource
 from op_tcg.backend.models.input import MetaFormat
+from op_tcg.backend.models.leader import LeaderElo
 
 
 class MatchResult(IntEnum):
@@ -41,15 +41,6 @@ class BQMatches(BaseModel):
     def to_dataframe(self) -> pd.DataFrame:
         return pd.DataFrame([r.dict() for r in self.matches])
 
-class LeaderElo(BQTableBaseModel):
-    _dataset_id: str = BQDataset.MATCHES
-    meta_format: MetaFormat = Field(description="Meta until or in which the elo is calculated", primary_key=True)
-    leader_id: str = Field(description="The op tcg leader id e.g. OP03-099", primary_key=True)
-    only_official: bool = Field(default=False, description="Whether the matches are only originated from "
-                                                           "official tournaments", primary_key=True)
-    elo: int = Field(description="Elo rating of leader until a certain time/ meta format")
-    start_date: date = Field(description="Date in which the elo calculation started")
-    end_date: date = Field(description="Date in which the elo calculation ended")
 
 class BQLeaderElos(BaseModel):
     elo_ratings: list[LeaderElo]
