+++ op_tcg/frontend/Leaderboard.py
@@ -7,8 +7,8 @@ from uuid import uuid4
 
 from op_tcg.backend.etl.load import bq_insert_rows, get_or_create_table
 from op_tcg.backend.models.input import MetaFormat
-from op_tcg.backend.models.leader import OPTcgColor, TournamentWinner
-from op_tcg.backend.models.matches import LeaderElo, BQLeaderElos, Match, MatchResult
+from op_tcg.backend.models.leader import OPTcgColor, TournamentWinner, LeaderElo
+from op_tcg.backend.models.matches import BQLeaderElos, Match, MatchResult
 from op_tcg.backend.models.bq_enums import BQDataset
 from op_tcg.frontend.utils.session import get_session_id
 from op_tcg.frontend.sidebar import display_meta_select, display_only_official_toggle, display_release_meta_select, \
