+++ pyproject.toml
@@ -21,6 +21,7 @@ requests = "^2.31.0"
 beautifulsoup4 = "^4.12.3"
 st-theme = "^1.2.2"
 pandera = "^0.19.3"
+tqdm = "^4.66.4"
 
 
 [tool.poetry.group.crawler.dependencies]
